Used subgraph provider for token distro
